pyproject.toml
[project]
name = "multidocchat"
version = "1.0.0"
description = "A production-grade multi-document RAG chatbot built with FastAPI, LangChain, FAISS, and deployed to AWS ECS Fargate"
authors = [
  {name = "Vikash", email = "vik.manit.19@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "docx2txt==0.9",
    "faiss-cpu>=1.12.0",
    "fastapi==0.115.6",
    "ipykernel==6.30.0",
    "jinja2==3.1.4",
    "langchain==0.3.27",
    "langchain-community==0.3.27",
    "langchain-core==0.3.72",
    "langchain-google-genai==2.1.8",
    "langchain-groq==0.3.6",
    "langchain-openai==0.2.10",
    "langsmith>=0.4.28",
    "openai>=1.107.3",
    "pandas>=2.3.3",
    "pypdf>=4.0.0",
    "python-dotenv==1.1.1",
    "python-multipart==0.0.20",
    "structlog>=25.4.0",
    "uvicorn==0.32.1",
    "boto3>=1.35.0",
    "langchain-pinecone==0.2.0",
    "sentence-transformers>=3.0.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
]
